Fix fade-during-fade behavior.
Previously, it would restart from the beginning of the fade, going to the new fade - so if you were going from red to blue, and then asked for black, it would jump back to red, and fade to black.  Jarring, if triggered.

The new behavior is better.  If a fade is running, it sets the new target, but extends the existing fade time.

So, if you have a 5 second fade from red to blue running, and then request a 5 second fade to black, 3 seconds in, you get a 7 second fade from (midway between red and blue) to black.

If you want to change it to not add the time on, this is pretty easy, but it ensures smooth fades, and more or less obeys the spirit of the fade request.

Our use case is fading from 100% to 0%, 10% at a time, for Good Friday services, and maintaining the smooth fade is important there.  Otherwise, this doesn't get tripped.